spliting out anchorItem, AddAnchorDialog and roundPosition
